Updates following February Tahoe Offsite
At the offsite this February, we discussed how our company mission does
not talk about what we're bringing to the world via our products, in
addition to our culture. This makes it more difficult for us to
communicate to others what we're building and selling. The main goal of
this revision is to center our desired cultural impact under the Values
section, and to make the Mission Statement about the product vision.

We're also adding a couple of new values as they are things we're
already doing which we want to preserve as we grow the team this year,
and have moved the current company goals to their own section.